Move DNS validation into models module
The models module should not depend on (an outdated version of) the main
module.
